edit question issue resolved

+++ src/pages/Location.js
@@ -10,6 +10,22 @@ export default function Location() {
   const [data, setData] = useState([]);
   const [perPage, setPerPage] = useState(10);
   const [currentPage, setCurrentPage] = useState(1);
+  const [showForm, setShowForm] = useState(false);
+  const [states, setStates] = useState([]);
+  const [modalHeading, setModalHeading] = useState("");
+  const [isFormEditCategory, setIsFormEditCategory] = useState(false);
+  const [editFormFields, setEditFormFields] = useState([]);
+  const [stateIdSelectedForEdit, setStateIdSelectedForEdit] = useState("");
+
+  const formFields = [
+    {
+      name: "state",
+      label: "State",
+      type: "dropdown",
+      options: states,
+    },
+    { name: "name", label: "Location", type: "text" },
+  ];
 
   const columns = [
     {
@@ -33,7 +49,12 @@ export default function Location() {
           <MdModeEditOutline
             color="#444"
             size={20}
-            onClick={() => handleEdit()}
+            onClick={() => {
+              setModalHeading("Edit Location");
+              setIsFormEditCategory(true);
+              handleEditFormFields(row);
+              handleToggleModal();
+            }}
             cursor="pointer"
           />
           <MdDelete
@@ -80,6 +101,29 @@ export default function Location() {
     getLocation();
   }, []);
 
+  const getStates = async () => {
+    // setLoading(true);
+
+    try {
+      const token = LocalStorage.getData("token");
+      const stateData = await http.get(`state/`, {
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+      });
+
+      setStates(stateData);
+    } catch (error) {
+      console.log(JSON.stringify(error), 25);
+    } finally {
+      //   setLoading(false);
+    }
+  };
+
+  useEffect(() => {
+    getStates();
+  }, []);
+
   const deleteLocation = async (id) => {
     try {
       const token = LocalStorage.getData("token");
@@ -93,16 +137,88 @@ export default function Location() {
     }
   };
 
-  const handleAdd = () => {
-    console.log("add");
+  const addLocation = async (locationData) => {
+    try {
+      const token = LocalStorage.getData("token");
+      const url = "location/";
+      const config = {
+        headers: {
+          " Authorization": `Bearer ${token}`,
+        },
+      };
+
+      const data = {
+        name: locationData.name,
+        state: { id: locationData.state },
+      };
+
+      const resp = await http.post(url, data, config);
+      console.log(resp);
+    } catch (error) {
+      console.log(JSON.stringify(error), 245);
+    } finally {
+      //   setLoading(false);
+    }
   };
 
-  const handleEdit = () => {
-    console.log("edit");
+  const handleAdd = async (event, values) => {
+    event.preventDefault();
+    console.log("hi", values);
+    await addLocation(values);
+    setShowForm(false);
+    setModalHeading("");
+  };
+
+  const updateLocation = async (locationData, id) => {
+    try {
+      const token = LocalStorage.getData("token");
+      const url = `location/${id}/`;
+      const config = {
+        headers: {
+          " Authorization": `Bearer ${token}`,
+        },
+      };
+
+      const data = {
+        state: { id: locationData.state.id },
+        name: locationData.name,
+      };
+
+      console.log(data);
+      const resp = await http.put(url, data, config);
+      console.log(resp);
+    } catch (error) {
+      console.log(JSON.stringify(error), 245);
+    } finally {
+      //   setLoading(false);
+    }
+  };
+
+  const handleEdit = async (event, values) => {
+    event.preventDefault();
+    console.log("hi", values);
+    await updateLocation(values, stateIdSelectedForEdit);
+    setShowForm(false);
+    setModalHeading("");
+  };
+
+  const handleEditFormFields = (rowData) => {
+    const fields = formFields.map((field) => {
+      return {
+        ...field,
+        defaultValue: rowData[field.name],
+      };
+    });
+    console.log(rowData, fields);
+    setStateIdSelectedForEdit(rowData.id);
+    setEditFormFields(fields);
   };
 
   const handleDelete = async (id) => {
-    await deleteLocation(id);
+    const shouldDelete = window.confirm("Are you sure you want to delete it ?");
+    if (shouldDelete) {
+      await deleteLocation(id);
+    }
   };
   const handlePageChange = async (page) => {
     setCurrentPage(page);
@@ -111,13 +227,19 @@ export default function Location() {
   const handlePerRowsChange = async (newPerPage) => {
     setPerPage(newPerPage);
   };
+  const handleToggleModal = () => {
+    setShowForm(!showForm);
+    isFormEditCategory && setIsFormEditCategory(false);
+  };
+
   return (
     <div>
       <NavigationBar />
       <div className="d-flex">
         <SidebarMenu />
         <MainContent
-          title="Area"
+          title="Location"
+          modalHeading={modalHeading}
           data={data}
           columns={columns}
           isRemote={false}
@@ -125,6 +247,13 @@ export default function Location() {
           handleDelete={handleDelete}
           handlePageChange={handlePageChange}
           handlePerRowsChange={handlePerRowsChange}
+          handleToggleModal={handleToggleModal}
+          isAddFormVisible={true}
+          showForm={showForm}
+          formFields={isFormEditCategory ? editFormFields : formFields}
+          setModalHeading={setModalHeading}
+          formSubmitHandler={isFormEditCategory ? handleEdit : handleAdd}
+          isFormEditCategory={isFormEditCategory}
         />
       </div>
     </div>
